Fix the specification of Ruby version
From Heroku's documentation:

As of bundler 1.12+ you can use a version specifier in your Ruby version. For
example if you want to deploy using 2.3.0 but some members of your team want to
use 2.3.1 you could allow this by specifying:

ruby "~> 2.3.0"

This is saying that any version of 2.3.x where x is greater than or equal to 0
are valid. For this to work on Heroku, you must specify the full version with
all three digits. For example neither ~> 2.3 or ~> 2.3.x are valid version
specifiers on Heroku.